# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hohaya API",
        "description": "API documentation for Hohaya platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Statistiques globales du dashboard admin",
                "operationId": "a4e953abeda97cdeaa1c2af786aea1f3",
                "responses": {
                    "200": {
                        "description": "Statistiques globales"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/revenus/mensuel": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Revenus mensuels sur les 12 derniers mois",
                "operationId": "c8aedf19608fd557c222ee41d332b9e7",
                "responses": {
                    "200": {
                        "description": "Revenus par mois"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste de tous les utilisateurs",
                "operationId": "8734f0eb21eaec99b614d902ba09f030",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filtrer par rôle",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "owner",
                                "seeker",
                                "admin"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom, email ou téléphone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_verified",
                        "in": "query",
                        "description": "Filtrer par statut de vérification",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "deleted",
                        "in": "query",
                        "description": "Afficher uniquement les comptes supprimés",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}/verify": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Vérifier ou retirer la vérification d'un propriétaire",
                "operationId": "14dba836d0c4ad63dee90222114d1a3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_verified"
                                ],
                                "properties": {
                                    "is_verified": {
                                        "description": "true = vérifier, false = retirer la vérification",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vérification mise à jour"
                    },
                    "422": {
                        "description": "Utilisateur non propriétaire"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}/suspension": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Suspendre ou réactiver un compte utilisateur",
                "operationId": "f2bb2ce1610dcb16da2935232b491101",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "suspendre"
                                ],
                                "properties": {
                                    "suspendre": {
                                        "description": "true = suspendre le compte, false = réactiver",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte suspendu ou réactivé"
                    },
                    "422": {
                        "description": "Utilisateur introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/properties/pending": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Annonces en attente de modération",
                "operationId": "cbd78c065af53441afd6d2409864098e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Annonces à modérer"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/export/paiements": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Exporter les paiements en CSV",
                "operationId": "6bf916e6b3d82f1679febe821469b8cf",
                "responses": {
                    "200": {
                        "description": "Fichier CSV des paiements"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/export/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Exporter les utilisateurs en CSV",
                "operationId": "a2e10752344944b1099680b20731d000",
                "responses": {
                    "200": {
                        "description": "Fichier CSV des utilisateurs"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}